move reports models to model module
