--- upload-server/internal/models/psreport.go
@@ -1,8 +1,0 @@
-package models
-
-// TODO: line up with PS reports
-type PsReport struct {
-	Service string
-	Stage   string
-	Report  string
-} //.PsReport
